+++ src/ConfigProvider.php
@@ -1,13 +1,12 @@
 <?php
+
 /**
- * @see https://github.com/zendframework/zend-expressive-authentication-session
- *     for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license https://github.com/zendframework/zend-expressive-authentication-session/blob/master/LICENSE.md
- *     New BSD License
+ * @see       https://github.com/mezzio/mezzio-authentication-session for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authentication-session/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authentication-session/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Expressive\Authentication\Session;
+namespace Mezzio\Authentication\Session;
 
 class ConfigProvider
 {
@@ -31,6 +30,10 @@ class ConfigProvider
     public function getDependencies() : array
     {
         return [
+            // Legacy Zend Framework aliases
+            'aliases' => [
+                \Zend\Expressive\Authentication\Session\PhpSession::class => PhpSession::class,
+            ],
             'factories' => [
                 PhpSession::class => PhpSessionFactory::class,
             ],