+++ CHANGELOG.md
@@ -22,12 +22,12 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#1](https://github.com/zendframework/zend-expressive-authentication-session/pull/1)
+- [zendframework/zend-expressive-authentication-session#1](https://github.com/zendframework/zend-expressive-authentication-session/pull/1)
   fixes how the `PhpSession` adapter both stores user details in the session,
-  and retrieves them. Since zend-expressive-session does not allow object
+  and retrieves them. Since mezzio-session does not allow object
   serialization, the class now stores the username and role in the session, and
   then populates an anonymous class implementing
-  `Zend\Expressive\Authentication\UserInterface` with the values on subsequent
+  `Mezzio\Authentication\UserInterface` with the values on subsequent
   requests.
 
 ## 0.1.0 - 2017-11-09